Add files via upload
This is what I have, I went through and worked on comments and that version of the file got deleted. I am planning on working on some of the sorting code and adding an option for choosing a local file instead of using the gspread option. I am also working on the README to explain how to use the program - note to reviewer